+++ src/main/java/ca/vastier/depense/services/ReceiptService.java
@@ -6,5 +6,5 @@ import ca.vastier.depense.web.dto.ReceiptDto;
 public interface ReceiptService
 {
 	ReceiptDto createReceipt(ReceiptDto receipt);
-	ReceiptDto findReceiptById(long id);
+	ReceiptDto findReceiptById(int id);
 }
